Add Filip's dummy sim, R project and expand gitignore
Filip's simulation of dummy data attemps to create data according to the model in equation 1 on page 17.

Created an R project in RStudio which created a .Rproj file that is recommended to include in git since it handles some settings that can be good to keep the same. Might be problem with the git-setup section in it?

Added .idea files to gitignore. Those are for managing the settings in PyCharm that Sebastian is using.