biclust takes cluster log
biclust now takes the entire cluster
history log so it can add to it. This
is so we can plot it later and understand
what has happened with the allocation
from start until end.

An alternative is to not send it
along and instead add the two cluster history
logs together after.

Also fixed some comments, prints and
naming.