Change cluster allocation
There is now an option, as standard,
to use mean metric per cell cluster
instead of min. The builtin aggregate
function is slow, so should be switched
for a faster package.

Added a variable to control disturbed
percentage of cells in initial clustering.

r2 is now adjuster for foreign cell clusters
in order to punish them.