Fix cell cluster history plot
The cell cluster history didn't keep track
of individual cells. Leading to when a cell
got merged with other cells it's position in
the blob wasn't related to anything.

This Sankey diagram should fix it. At least
it looks like it when running toy data, and
the input data it wants contains node to node
info.